+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [[Unreleased]]
+- @xrplf/eslint-config allows peers with typescript 3 or 4
+- @xrplf/eslint-config skips blank lines and comments in
+  `max-lines` and `max-lines-per-function` rules
+
 
 ## 1.2.2
 First version that is correctly tagged with this CHANGELOG. Previous releases